% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{cor.trial}
\alias{cor.trial}
\title{correlation}
\usage{
cor.trial(y, x, args)
}
\arguments{
\item{y}{response}

\item{x}{covariates}

\item{args}{see ...}
}
\value{
list of vectors
}
\description{
this function ...
}
\examples{
NA
}
