% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{dims.trial}
\alias{dims.trial}
\title{dimensionality}
\usage{
dims.trial(y, X, args = NULL)
}
\arguments{
\item{y}{to do}

\item{X}{to do}

\item{args}{to do}
}
\value{
to do
}
\description{
this function ...
}
\examples{
NA
}
