% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.args}
\alias{.args}
\title{Arguments}
\usage{
.args(...)
}
\arguments{
\item{...}{Arguments supplied to \code{\link{palasso}},
other than \code{y}, \code{X} and \code{max}.}
}
\value{
Returns the arguments as a list, including default values
for missing arguments.
}
\description{
Checks the validity of the provided arguments.
}
\examples{
NA

}
