% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.cor}
\alias{.cor}
\title{Correlation}
\usage{
.cor(y, x, args)
}
\arguments{
\item{y}{vector of length \eqn{n}}

\item{x}{matrix with \eqn{n} rows and \eqn{p} columns}

\item{args}{options for paired lasso\strong{:}
list of arguments
(output from \link{.dims} and \link{.args})}
}
\value{
list of vectors
}
\description{
Calculates the correlation between the response and the covariates.
Shrinks the correlation coefficients for each covariate set separately.
}
\examples{
NA

}
\keyword{internal}
