% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.cv}
\alias{.cv}
\title{Cross-validation}
\usage{
.cv(y, x, foldid, lambda, args)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{x}{covariates\strong{:}
matrix with \eqn{n} rows (samples)
and \eqn{k * p} columns (variables)}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n},
with entries from \eqn{1} to \code{nfolds}}

\item{lambda}{lambda sequence\strong{:}
vector of decreasing positive values}

\item{args}{options for paired lasso\strong{:}
list of arguments
(output from \link{.dims} and \link{.args})}
}
\value{
Returns matrix of predicted values (except "cox")
}
\description{
Repeatedly leaves out samples, and predicts their response.
}
\examples{
NA

}
