% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.dims}
\alias{.dims}
\title{Dimensionality}
\usage{
.dims(y, X, args = NULL)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
list of matrices,
each with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{args}{options for paired lasso\strong{:}
list of arguments
(output from \link{.dims} and \link{.args})}
}
\value{
The function \code{.dims} extracts the dimensionality.
It returns the numbers of samples,
covariate pairs and covariate sets.
It also returns the number of weighting schemes,
and the names of these weighting schemes.
}
\description{
This function extracts the dimensions.
}
\examples{
NA

}
