% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.extract}
\alias{.extract}
\title{Extraction}
\usage{
.extract(fit, lambda, cvm, type.measure)
}
\arguments{
\item{fit}{matrix with one row for each sample
("gaussian", "binomial" and "poisson"),
or one row for each fold (only "cox"),
and one column for each \code{lambda}
(output from \link{.fit})}

\item{lambda}{lambda sequence\strong{:}
vector of decreasing positive values}

\item{cvm}{mean cross-validated loss\strong{:}
vector of same length as \code{lambda}
(output from \link{.loss})}

\item{type.measure}{...
loss function\strong{:}
character "deviance", "mse", "mae", "class", or "auc"}
}
\description{
Extracts \code{cv.glmnet}-like object.
}
\examples{
NA

}
\keyword{internal}
