% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.fit}
\alias{.fit}
\title{Model bag}
\usage{
.fit(y, x, args)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{x}{covariates\strong{:}
matrix with \eqn{n} rows (samples)
and \eqn{k * p} columns (variables)}

\item{args}{options for paired lasso\strong{:}
list of arguments
(output from \link{.dims} and \link{.args})}
}
\value{
list of \code{glmnet}-like objects
}
\description{
Fits all models from the chosen bag.
}
\examples{
NA
}
