% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.folds}
\alias{.folds}
\title{Cross-validation folds}
\usage{
.folds(y, nfolds, foldid = NULL)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{nfolds}{number of folds\strong{:}
positive integer
(\eqn{>= 10} recommended)}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n},
with entries from \eqn{1} to \code{nfolds}}
}
\value{
Returns the fold identifiers.
}
\description{
Assigns samples to cross-validation folds,
balancing the folds in the case of a binary or survival response.
}
\examples{
NA

}
\keyword{internal}
