% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.loss}
\alias{.loss}
\title{Cross-validation loss}
\usage{
.loss(y, fit, family, type.measure, foldid = NULL)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{fit}{matrix with one row for each sample
("gaussian", "binomial" and "poisson"),
or one row for each fold (only "cox"),
and one column for each \code{lambda}
(output from \link{.fit})}

\item{family}{model family\strong{:}
character "gaussian", "binomial", "poisson", or "cox"}

\item{type.measure}{...
loss function\strong{:}
character "deviance", "mse", "mae", "class", or "auc"}

\item{foldid}{fold identifiers\strong{:}
vector of length \eqn{n},
with entries from \eqn{1} to \code{nfolds}}
}
\value{
Returns list of vectors, one for each model.
}
\description{
Calculates mean cross-validated loss
}
\examples{
NA

}
\keyword{internal}
