% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{.weight}
\alias{.weight}
\title{Weighting schemes}
\usage{
.weight(cor, args)
}
\arguments{
\item{cor}{correlation coefficients\strong{:}
list of \eqn{k} vectors of length \eqn{p}
(one vector for each covariate set with
one entry for each covariate)}

\item{args}{options for paired lasso\strong{:}
list of arguments
(output from \link{.dims} and \link{.args})}
}
\value{
list of named vectors (one for each weighting scheme)
}
\description{
Calculates the weighting schemes.
}
\examples{
NA

}
