% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extra}
\alias{extra}
\alias{.prepare}
\alias{.simulate}
\alias{.predict}
\alias{.select}
\title{extra}
\usage{
.prepare(X, cutoff = NULL)

.simulate(x, effects)

.predict(y, X, pmax = NULL, nfolds.ext = 5, nfolds.int = 5)

.select(y, X, index, pmax = 10, nfolds = 5)
}
\arguments{
\item{X}{covariates\strong{:}
matrix with \eqn{n} rows and \eqn{p} columns}

\item{x}{covariates\strong{:}
list of length \eqn{k},
including matrices with \eqn{n} rows and \eqn{p} columns}

\item{effects}{number of causal covariates\strong{:}
vector of length \eqn{k}}

\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{nfolds.ext}{number of external folds}

\item{index}{indices of causal covariates\strong{:}
list of length \eqn{k},
including vectors}

\item{...}{arguments for \link[palasso]{palasso}}
}
\value{
to do
}
\description{
generic functions
}
\details{
By default, the function \code{predict} returns
the linear predictor (\code{type="link"}).
Consider predicting the response (\code{type="response"}).
}
\examples{
set.seed(1)
n <- 20; p <- 50
X <- matrix(rpois(n*p,lambda=4),nrow=n,ncol=p)
x <- .prepare(X)
y <- .simulate(x,effects=c(1,2))
.predict(y,x)
.select(y,x,attributes(y))
1+1
1+1
1+1
1+1
}
\seealso{
Use \link[palasso]{palasso} to fit the paired lasso.
}
\keyword{internal}
