% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract.trial}
\alias{extract.trial}
\title{extract}
\usage{
extract.trial(fit, lambda, cvm, type.measure)
}
\arguments{
\item{fit}{...}

\item{lambda}{...}

\item{cvm}{...}

\item{type.measure}{...}
}
\value{
this ...
}
\description{
this function ...
}
\details{
to do
}
\examples{
NA
}
