% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{fit.trial}
\alias{fit.trial}
\title{fit models}
\usage{
fit.trial(y, X, args)
}
\arguments{
\item{y}{response}

\item{X}{covariates}

\item{args}{see ...}
}
\value{
list of glmnet-like objects
}
\description{
this function ...
}
\examples{
NA
}
