% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{folds.trial}
\alias{folds.trial}
\title{to do}
\usage{
folds.trial(y, nfolds, foldid)
}
\arguments{
\item{y}{response}

\item{nfolds}{...}

\item{foldid}{...}
}
\value{
to do
}
\description{
to do
}
\examples{
NA
}
