% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mtext}
\alias{mtext}
\title{Split}
\usage{
mtext(text, unit = FALSE, side = 1)
}
\arguments{
\item{text}{character vector}

\item{unit}{logical}

\item{side}{to do}
}
\value{
to do
}
\description{
generic functions
}
\details{
to do
}
\keyword{internal}
