% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{palasso}
\alias{palasso}
\alias{palasso-package}
\title{Paired lasso regression}
\usage{
palasso(y = y, X = X, max = 10, ...)
}
\arguments{
\item{y}{trial}

\item{X}{trial}

\item{max}{trial}

\item{...}{trial}
}
\value{
trial
}
\description{
trial
}
\details{
trial
}
\examples{
if(FALSE){
    start <- Sys.time()
    for(i in 1:100){
        cat("i =",i,"\\n")
        set.seed(i)
        n <- 50; p <- 20
        y <- rbinom(n=n,size=1,prob=0.5)
        X <- lapply(1:2,function(x) matrix(rnorm(n*p),nrow=n,ncol=p))
        object <- palasso(y=y,X=X,family="binomial")
        a <- predict(object,newdata=X)
        b <- coef(object)
        c <- weights(object)
        d <- fitted(object)
        e <- residuals(object)
        f <- deviance(object)
        g <- logLik(object)
        #h <- summary(object)
    }
    end <- Sys.time()
    end-start
}

}
\references{
trial
}
