% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.matrix}
\alias{plot.matrix}
\title{plot matrix}
\usage{
\method{plot}{matrix}(x, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{to do}
}
\value{
to do
}
\description{
generic functions
}
\details{
to do
}
\seealso{
Use \link[palasso]{palasso} to fit the pairwise-adaptive lasso.
}
\keyword{plots}
