% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatter}
\alias{scatter}
\title{Set comparison}
\usage{
scatter(x, y, p = 0.95, ...)
}
\arguments{
\item{x}{vector of length \eqn{n}}

\item{y}{vector of length \eqn{n}}

\item{p}{confidence interval\strong{:}
numeric between \eqn{0} and \eqn{1}}

\item{...}{Graphical arguments}
}
\value{
This function returns a plot.
}
\description{
The function \code{scatter} plots the difference
between two sets of measurements.
}
\examples{
x <- runif(100)
y <- runif(100)
scatter(x,y)

}
\keyword{internal}
