% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{sim.trial}
\alias{sim.trial}
\title{to do}
\usage{
sim.trial(family, n = 200, p = 150)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}
}
\value{
to do
}
\description{
to do
}
\examples{
NA
}
