% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toydata}
\format{A list of numeric vectors and numeric matrices.}
\usage{
data(toydata)
}
\value{
All entries are numeric.
}
\description{
This dataset allows you to reproduce
the examples shown in the vignette.
}
\keyword{internal}
