% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{weight.trial}
\alias{weight.trial}
\title{compute weights}
\usage{
weight.trial(cor, args)
}
\arguments{
\item{cor}{see ...}

\item{args}{see ...}
}
\value{
list of vectors
}
\description{
this function ...
}
\examples{
NA
}
