% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{cv.trial}
\alias{cv.trial}
\title{cross-validation}
\usage{
cv.trial(y, x, foldid, lambda, args)
}
\arguments{
\item{y}{response}

\item{x}{covariates}

\item{foldid}{vector}

\item{lambda}{vector}

\item{args}{see ...}
}
\value{
to do
}
\description{
this function
}
\examples{
NA
}
