% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{extra}
\alias{extra}
\alias{.prepare}
\alias{.simulate}
\alias{.predict}
\alias{.select}
\title{Hidden functions (analysis)}
\usage{
.prepare(X, cutoff = NULL, quantile = NULL, scale = TRUE)

.simulate(x, effects)

.predict(y, X, nfolds.ext = 5, nfolds.int = 5, ...)

.select(y, X, index, family = "binomial", nfolds = 5, ...)
}
\arguments{
\item{X}{covariates\strong{:}
matrix with \eqn{n} rows and \eqn{p} columns}

\item{x}{covariates\strong{:}
list of length \eqn{k},
including matrices with \eqn{n} rows and \eqn{p} columns}

\item{effects}{number of causal covariates\strong{:}
vector of length \eqn{k}}

\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{nfolds.ext}{number of external folds}

\item{...}{arguments for \link[palasso]{palasso}}

\item{index}{indices of causal covariates\strong{:}
list of length \eqn{k},
including vectors}

\item{trial}{development option}
}
\value{
to do
}
\description{
Functions for the \code{palasso} manuscript.
}
\details{
The function \code{.prepare} pre-processes sequencing data. It removes
features with a low total abundance, adjusts for different library sizes,
binarises the counts with cutoff zero or takes the Anscombe transform, and
scales all covariates to mean zero and unit variance.
(If the cutoff differed from zero, we would have to first normalise,
then binarise, and finally filter the raw counts.)

The function \code{.simulate} simulates a response. Exploiting an
experimental covariate matrix, it allows for different numbers of non-zero
coefficients for X and Z.

The function \code{.predict} estimates the predictive performance of
different lasso models (standard X and/or Z, adaptive X and/or Z, paired
X and Z). Minimising the deviance, it returns the deviance and the area
under the curve. Currently, only the binomial family is implemented.

The function \code{.select} estimates the selective performance of different
lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z).
Limiting the number of covariates to \eqn{10}, it returns the number of
selected covariates, and the number of correctly selected covariates.
}
\examples{
set.seed(1)
n <- 50; p <- 100
X <- matrix(rpois(n*p,lambda=4),nrow=n,ncol=p)
x <- palasso:::.prepare(X)
y <- palasso:::.simulate(x,effects=c(1,2))
palasso:::.predict(y,x)
palasso:::.select(y,x,attributes(y))




}
\seealso{
Use \link[palasso]{palasso} to fit the paired lasso.
}
\keyword{internal}
