% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{loss.trial}
\alias{loss.trial}
\title{to do}
\usage{
loss.trial(y, fit, family, type.measure, foldid = NULL)
}
\arguments{
\item{y}{response}

\item{fit}{see ..}

\item{family}{...}

\item{type.measure}{...}
}
\value{
to do
}
\description{
to do
}
\examples{
NA
}
