% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{methods}
\alias{methods}
\alias{subset.palasso}
\alias{predict.palasso}
\alias{coef.palasso}
\alias{fitted.palasso}
\alias{residuals.palasso}
\alias{deviance.palasso}
\alias{logLik.palasso}
\alias{weights.palasso}
\alias{summary.palasso}
\alias{print.palasso}
\title{Methods for class "palasso"}
\usage{
\method{subset}{palasso}(x, model = "paired", ...)

\method{predict}{palasso}(object, newdata, s = "lambda.min",
  model = "paired", ...)

\method{coef}{palasso}(object, s = "lambda.min", model = "paired", ...)

\method{fitted}{palasso}(object, s = "lambda.min", model = "paired", ...)

\method{residuals}{palasso}(object, s = "lambda.min", model = "paired", ...)

\method{deviance}{palasso}(object, model = "paired", ...)

\method{logLik}{palasso}(object, s = NULL, model = "paired", ...)

\method{weights}{palasso}(object, model = "paired", ...)

\method{summary}{palasso}(object, model = "paired", ...)

\method{print}{palasso}(x, ...)
}
\arguments{
\item{model}{character "paired" (or one of \code{names(x)})}

\item{...}{further arguments for \link[glmnet]{predict.cv.glmnet},
\link[glmnet]{coef.cv.glmnet}, or \link[glmnet]{deviance.glmnet}}

\item{object, x}{\link[palasso]{palasso} object}

\item{newdata}{covariates\strong{:}
list of matrices with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{s}{penalty parameter\strong{:}
character "lambda.min" or "lambda.1se",
or positive numeric}
}
\value{
to do
}
\description{
generic functions
}
\details{
By default, the function \code{predict} returns
the linear predictor (\code{type="link"}).
Consider predicting the response (\code{type="response"}).
}
\seealso{
Use \link[palasso]{palasso} to fit the paired lasso.
}
