% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{methods}
\alias{methods}
\alias{subset.palasso}
\alias{predict.palasso}
\alias{coef.palasso}
\alias{fitted.palasso}
\alias{deviance.palasso}
\alias{weights.palasso}
\alias{summary.palasso}
\title{Methods for class "palasso"}
\usage{
\method{subset}{palasso}(x, model = "paired", ...)

\method{predict}{palasso}(object, newdata, s = "lambda.min",
  model = "paired", ...)

\method{coef}{palasso}(object, s = "lambda.min", model = "paired", ...)

\method{fitted}{palasso}(object, s = "lambda.min", model = "paired", ...)

\method{deviance}{palasso}(object, model = "paired", ...)

\method{weights}{palasso}(object, model = "paired", ...)

\method{summary}{palasso}(object, model = "paired", ...)
}
\arguments{
\item{model}{character "paired"}

\item{...}{further arguments for \link[glmnet]{predict.cv.glmnet},
\link[glmnet]{coef.cv.glmnet}, or \link[glmnet]{deviance.glmnet}
# model character "paired", "standard", "adaptive"
# groups character "X" or "Z", or NULL (all groups)}

\item{object, x}{\link[palasso]{palasso} object}

\item{newdata}{covariates\strong{:}
list of matrices with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{s}{penalty parameter\strong{:}
character "lambda.min" or "lambda.1se",
or positive numeric}
}
\value{
to do
}
\description{
generic functions
}
\details{
to do
}
\seealso{
Use \link[palasso]{palasso} to fit the pairwise-adaptive lasso.
}
