% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palasso}
\alias{palasso}
\title{Pairwise-adaptive lasso}
\usage{
palasso(y, X, trial = FALSE, ...)
}
\arguments{
\item{y}{response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
list of matrices with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{trial}{development mode\strong{:}
logical (temporary argument)}

\item{...}{further arguments for \code{\link[glmnet]{cv.glmnet}}
or \code{\link[glmnet]{glmnet}}}
}
\value{
This function returns an object of class \code{palasso}.
}
\description{
The function \code{palasso} cross-validates
the pairwise-adaptive lasso.
Use this regression technique if
the covariates are numerous and occur in pairs.
}
\details{
Let \code{x} denote one entry of the list \code{X}.
See \link[glmnet]{glmnet} for alternative
specifications of \code{y} and \code{x}.
}
\examples{
set.seed(1)
n <- 100; p <- 200
y <- rbinom(n=n,size=1,prob=0.5)
X <- lapply(1:2,function(x) matrix(rnorm(n*p),nrow=n,ncol=p))
fit <- palasso(y=y,X=X,family="binomial",trial=TRUE)

}
\seealso{
Available \code{\link[palasso]{methods}} for class \code{palasso}:
\code{\link[=predict.palasso]{predict}},
\code{\link[=coef.palasso]{coef}},
\code{\link[=fitted.palasso]{fitted}},
\code{\link[=weights.palasso]{weights}},
\code{\link[=deviance.palasso]{deviance}},
\code{\link[=summary.palasso]{summary}},
\code{\link[=subset.palasso]{subset}}
}
