% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plots}
\alias{plots}
\alias{plot_score}
\alias{plot_image}
\title{Plots for class "palasso"}
\usage{
plot_score(X, choice)

plot_image(X, margin = 2, text = TRUE)
}
\arguments{
\item{X, Y}{matrix with \eqn{n} rows and \eqn{p} columns}

\item{choice}{numeric between \eqn{1} and \eqn{p}}

\item{x, y}{vector}

\item{names}{vector of length \eqn{p}}

\item{groups}{vector of length \eqn{p}}

\item{...}{to do}
}
\value{
to do
}
\description{
generic functions
}
\details{
to do
}
\seealso{
Use \link[palasso]{palasso} to fit the pairwise-adaptive lasso.
}
